refactor(index): extract index file
