+++ src/cacheorm/__init__.py
@@ -1,6 +1,6 @@
 from .backends import *
 from .fields import *
-from .models import *
+from .model import *
 from .serializers import *
 
 __version__ = "0.0.1"
