+++ src/cacheorm/model.py
@@ -1,57 +1,9 @@
 import copy
-import itertools
 from collections import defaultdict
 
 from .fields import Field, FieldAccessor, IntegerField
+from .index import IndexManager
 from .types import with_metaclass
-
-
-class Index(object):
-    def __init__(self, model, fields, formatter=None):
-        self.model = model
-        self.fields = fields
-        self.field_names = {field.name for field in fields}
-        if formatter is None:
-            formatter = self._default_formatter(model, fields)
-        self.formatter = formatter
-
-    def make_cache_key(self, **query):
-        missing_keys = self.field_names - set(query.keys())
-        if missing_keys:
-            raise KeyError("missing index keys %s in query" % missing_keys)
-        values = tuple(field.cache_value(query[field.name]) for field in self.fields)
-        return self.formatter % values
-
-    @staticmethod
-    def _default_formatter(model, fields):
-        base = "m:%s:" % model._meta.name
-        field_parts = ":".join(itertools.chain(*[(f.name, "%s") for f in fields]))
-        return base + field_parts
-
-
-class PrimaryKeyIndex(Index):
-    def __init__(self, model, formatter=None):
-        super(PrimaryKeyIndex, self).__init__(
-            model, model._meta.get_primary_key_fields(), formatter
-        )
-
-
-class IndexManager(object):
-    def __init__(self, model):
-        self.model = model
-        self.indexes = {}
-
-    def _generate_indexes(self):
-        primary_key = self.model._meta.primary_key
-        self.indexes[primary_key] = PrimaryKeyIndex(
-            self.model, formatter=primary_key.index_formatter
-        )
-
-    def get_primary_key_index(self):
-        return self.get_index(self.model._meta.primary_key)
-
-    def get_index(self, field):
-        return self.indexes.get(field, None)
 
 
 class Metadata(object):
@@ -584,7 +536,7 @@ class ModelUpdate(Update):
 
 
 # NOTE(leosocy):
-#  1. IndexSelector 根据query，生成所有匹配的indexes，交给调用者决定如何使用索引。
+#  1. IndexMatcher 根据query，生成所有匹配的indexes，交给调用者决定如何使用索引。
 #  2. Where 根据index及传入的query，分离index string以及剩余的payload。其中index需要包括convert(cache_value)逻辑。
 #  3. Payloader to_cache, from_cache。
 #   - 根据传入的query，调用cache_value生成需要保存到cache backend的值。
